chore: Move Tooltip component to UI directory
