--- packages/webapp-libs/webapp-core/src/components/tooltip/index.ts
@@ -1,2 +1,0 @@
-export * from './tooltipContent';
-export * from './tooltip.component';
