--- packages/webapp-libs/webapp-core/src/components/tooltip/tooltip.component.tsx
@@ -1,9 +1,0 @@
-import * as TooltipPrimitive from '@radix-ui/react-tooltip';
-
-const TooltipProvider = TooltipPrimitive.Provider;
-
-const Tooltip = TooltipPrimitive.Root;
-
-const TooltipTrigger = TooltipPrimitive.Trigger;
-
-export { Tooltip, TooltipTrigger, TooltipProvider };
