--- packages/webapp-libs/webapp-core/src/components/tooltip/tooltipContent/index.ts
@@ -1,1 +1,0 @@
-export * from './TooltipContent.component';
