+++ packages/webapp-libs/webapp-core/src/components/ui/__stories__/tooltip.stories.tsx
@@ -1,7 +1,6 @@
 import { Meta, StoryFn, StoryObj } from '@storybook/react';
 
-import { Tooltip, TooltipProvider, TooltipTrigger } from './tooltip.component';
-import { TooltipContent } from './tooltipContent';
+import { Tooltip, TooltipContent, TooltipProvider, TooltipTrigger } from '../tooltip';
 
 const Template: StoryFn = () => {
   return (
